feat: Implement new chat UI with sidebar layout
- Move chat-view from lib to components with separate HTML/CSS files
- Create full-height fixed layout with sidebar and main content area
- Add sidebar with OpenAgents branding, search, and chat history
- Implement dark theme with monospace fonts
- Add floating header with user controls
- Create centered chat area with messages and input sections
- Remove blue focus ring from textarea and add autofocus
- Backup old implementations to chat-old and home-old routes
- Update all @effect/experimental packages to v0.49.2 to fix version conflicts

The new UI provides a clean, modern chat interface with proper separation of concerns using dedicated HTML and CSS files for better maintainability and syntax highlighting.
