+++ apps/example-psionic/package.json
@@ -10,7 +10,7 @@
   },
   "dependencies": {
     "@openagentsinc/psionic": "workspace:*",
-    "effect": "3.16.3",
+    "effect": "3.16.8",
     "elysia": "^1.3.5"
   },
   "devDependencies": {
