+++ apps/openagents.com/package.json
@@ -12,9 +12,9 @@
     "generate:llms-txt": "tsx ../../scripts/generate-llms-txt.ts"
   },
   "dependencies": {
-    "@effect/platform": "0.84.6",
-    "@effect/platform-bun": "^0.70.4",
-    "@effect/platform-node": "0.85.2",
+    "@effect/platform": "0.85.2",
+    "@effect/platform-bun": "^0.70.5",
+    "@effect/platform-node": "0.86.5",
     "@noble/curves": "1.2.0",
     "@noble/hashes": "1.3.1",
     "@openagentsinc/ai": "workspace:*",
@@ -24,7 +24,7 @@
     "@openagentsinc/sdk": "workspace:*",
     "@openagentsinc/ui": "workspace:*",
     "dotenv": "^16.4.7",
-    "effect": "3.16.3",
+    "effect": "3.16.8",
     "elysia": "^1.3.5",
     "mysql2": "^3.11.5"
   },
