+++ package.json
@@ -38,7 +38,7 @@
     "@typescript-eslint/eslint-plugin": "8.33.1",
     "@typescript-eslint/parser": "8.33.1",
     "babel-plugin-annotate-pure-calls": "0.4.0",
-    "effect": "3.16.3",
+    "effect": "3.16.8",
     "eslint": "9.28.0",
     "eslint-import-resolver-typescript": "4.4.2",
     "eslint-plugin-codegen": "0.31.0",
