+++ packages/autotest/package.json
@@ -34,9 +34,9 @@
     "orchestrate": "bun run src/orchestrate.ts"
   },
   "dependencies": {
-    "@effect/platform": "0.84.6",
-    "@effect/platform-bun": "^0.70.2",
-    "effect": "3.16.3",
+    "@effect/platform": "0.85.2",
+    "@effect/platform-bun": "^0.70.5",
+    "effect": "3.16.8",
     "effect-mcp": "^0.1.3",
     "puppeteer": "23.10.4"
   },
