+++ packages/container/package.json
@@ -19,10 +19,10 @@
   },
   "effect": {},
   "dependencies": {
-    "@effect/platform": "0.84.6",
-    "@effect/platform-node": "0.85.2",
+    "@effect/platform": "0.85.2",
+    "@effect/platform-node": "0.86.5",
     "@effect/schema": "0.75.5",
-    "effect": "3.16.3"
+    "effect": "3.16.8"
   },
   "devDependencies": {
     "@babel/cli": "7.27.2",
