+++ packages/nostr/package.json
@@ -28,7 +28,7 @@
     "coverage": "vitest --coverage"
   },
   "dependencies": {
-    "@effect/platform": "0.84.6",
+    "@effect/platform": "0.85.2",
     "@noble/ciphers": "^0.5.1",
     "@noble/curves": "1.8.1",
     "@noble/hashes": "1.3.3",
@@ -36,7 +36,7 @@
     "@scure/bip32": "^1.7.0",
     "@scure/bip39": "^1.6.0",
     "bech32": "^2.0.0",
-    "effect": "3.16.3",
+    "effect": "3.16.8",
     "ws": "8.18.0"
   },
   "devDependencies": {
