+++ packages/relay/package.json
@@ -14,7 +14,7 @@
     "directory": "dist"
   },
   "main": "./dist/cjs/index.js",
-  "module": "./dist/esm/index.js", 
+  "module": "./dist/esm/index.js",
   "types": "./dist/dts/index.d.ts",
   "exports": {
     ".": {
@@ -40,33 +40,33 @@
     "db:studio": "drizzle-kit studio"
   },
   "dependencies": {
-    "@effect/experimental": "^0.47.0",
-    "@effect/platform": "^0.83.0",
-    "@effect/sql": "^0.37.0",
-    "@effect/sql-mysql2": "^0.37.0", 
+    "@effect/experimental": "0.49.2",
+    "@effect/platform": "^0.85.2",
+    "@effect/sql": "^0.38.2",
+    "@effect/sql-mysql2": "^0.39.2",
     "@planetscale/database": "^1.19.0",
     "drizzle-orm": "^0.44.2",
-    "mysql2": "^3.11.5",
-    "effect": "3.16.3",
-    "elysia": "^1.3.5"
+    "effect": "3.16.8",
+    "elysia": "^1.3.5",
+    "mysql2": "^3.11.5"
   },
   "devDependencies": {
-    "@openagentsinc/nostr": "workspace:*",
-    "@openagentsinc/psionic": "workspace:*",
-    "drizzle-kit": "^0.30.1",
-    "@types/node": "^22.15.29",
-    "dotenv": "^16.4.7",
     "@babel/cli": "^7.25.12",
     "@babel/core": "^7.26.0",
     "@babel/plugin-transform-export-namespace-from": "^7.25.9",
     "@babel/plugin-transform-modules-commonjs": "^7.25.9",
-    "babel-plugin-annotate-pure-calls": "^0.4.0",
     "@effect/build-utils": "^0.7.8",
-    "@types/ws": "^8.5.14",
-    "ws": "^8.18.0",
+    "@effect/vitest": "^0.23.3",
     "@noble/curves": "^1.8.1",
     "@noble/hashes": "^1.3.3",
+    "@openagentsinc/nostr": "workspace:*",
+    "@openagentsinc/psionic": "workspace:*",
+    "@types/node": "^22.15.29",
+    "@types/ws": "^8.5.14",
+    "babel-plugin-annotate-pure-calls": "^0.4.0",
+    "dotenv": "^16.4.7",
+    "drizzle-kit": "^0.30.1",
     "vitest": "^3.2.1",
-    "@effect/vitest": "^0.23.3"
+    "ws": "^8.18.0"
   }
 }
