+++ packages/sdk/package.json
@@ -32,7 +32,7 @@
     "@openagentsinc/ai": "workspace:^",
     "@openagentsinc/container": "workspace:^",
     "@openagentsinc/nostr": "workspace:^",
-    "effect": "3.16.3"
+    "effect": "3.16.8"
   },
   "devDependencies": {
     "@babel/cli": "7.27.2",
